Extract test tasks from task.py
